+++ src/engine/LazyGroupBy.h
@@ -5,7 +5,7 @@
 #ifndef QLEVER_SRC_ENGINE_LAZYGROUPBY_H
 #define QLEVER_SRC_ENGINE_LAZYGROUPBY_H
 
-#include "engine/GroupBy.h"
+#include "engine/GroupByImpl.h"
 
 // Helper class to lazily compute the result of a group by operation. It makes
 // use of the hash map optimization to store the intermediate results of the
