+++ test/engine/LazyGroupByTest.cpp
@@ -8,7 +8,7 @@
 #include "../util/IndexTestHelpers.h"
 #include "./ValuesForTesting.h"
 #include "engine/GroupBy.h"
-#include "engine/GroupByWrapper.h"
+#include "engine/GroupByImpl.h"
 #include "engine/LazyGroupBy.h"
 #include "engine/sparqlExpressions/AggregateExpression.h"
 #include "engine/sparqlExpressions/GroupConcatExpression.h"
@@ -75,7 +75,7 @@ class LazyGroupByTest : public ::testing::Test {
 // _____________________________________________________________________________
 TEST_F(LazyGroupByTest, verifyEmptyGroupsAreAggregatedCorrectly) {
   IdTable resultTable{2, ua};
-  GroupBy::GroupBlock block{{0, I(7)}};
+  GroupByImpl::GroupBlock block{{0, I(7)}};
   IdTable idTable{1, ad_utility::makeAllocatorWithLimit<Id>(0_B)};
   auto evaluationContext = makeEvaluationContext(idTable);
 
@@ -95,7 +95,7 @@ TEST_F(LazyGroupByTest, verifyEmptyGroupsAreAggregatedCorrectly) {
 // _____________________________________________________________________________
 TEST_F(LazyGroupByTest, verifyGroupsAreAggregatedCorrectly) {
   IdTable resultTable{2, ua};
-  GroupBy::GroupBlock block{{0, I(7)}};
+  GroupByImpl::GroupBlock block{{0, I(7)}};
   IdTable idTable = makeIdTableFromVector({{2}, {3}, {5}, {7}}, I);
   auto evaluationContext = makeEvaluationContext(idTable);
 
@@ -120,7 +120,7 @@ TEST_F(LazyGroupByTest, verifyGroupsAreAggregatedCorrectly) {
 // _____________________________________________________________________________
 TEST_F(LazyGroupByTest, verifyCommitWorksWhenOriginalIdTableIsGone) {
   IdTable resultTable{2, ua};
-  GroupBy::GroupBlock block{{0, I(3)}};
+  GroupByImpl::GroupBlock block{{0, I(3)}};
   {
     IdTable idTable = makeIdTableFromVector({{2}, {3}, {5}, {7}}, I);
     auto evaluationContext = makeEvaluationContext(idTable);
